enhance Makefile and scripts: add swap file configuration options in Makefile, implement swap file management in setup, and improve DNS resolution script to conditionally execute based on environment variable; update install-docker-ce.sh to configure Docker daemon settings

+++ bin/fix-dns-resolv.sh
@@ -1,12 +1,22 @@
 # DNS
-sudo sed -i 's/^#DNSStubListener=.*/DNSStubListener=no/' /etc/systemd/resolved.conf
-# sudo systemctl restart systemd-resolved
-sudo systemctl disable --now systemd-resolved
-sudo rm -f /etc/resolv.conf
+## test if UPDATE_DNS_RESOLVERS is set to true in .env
+if [ "${UPDATE_DNS_RESOLVERS}" != "true" ] && [ "${UPDATE_DNS_RESOLVERS}" != "1" ]; then
+    echo "UPDATE_DNS_RESOLVERS is not set to true in .env, skipping DNS resolution fix"
+    exit 0
+fi
 
+sudo sed -i 's/^#DNSStubListener=.*/DNSStubListener=no/' /etc/systemd/resolved.conf
+sudo systemctl stop systemd-resolved
+sudo systemctl disable systemd-resolved
+sudo systemctl mask systemd-resolved
+sudo rm -f /etc/resolv.conf
 sudo tee /etc/resolv.conf << EOF > /dev/null
-nameserver ${NAMESERVER1}
-nameserver ${NAMESERVER2}
-nameserver ${NAMESERVER3}
-search ${INFRA_NAME} ${INFRA_NAME}.${INFRA_DOMAIN} ${SEARCH_DOMAIN}
+nameserver ${EXTERNAL_IP}
+nameserver ${GATEWAY_IP}
+nameserver 8.8.8.8
+search ${INFRA_NAME}.${INFRA_DOMAIN} ${INFRA_NAME}.${INFRA_DOMAIN/-/.} ${SEARCH_DOMAIN}
 EOF
+
+# Confirm the DNS configuration:
+cat /etc/resolv.conf
+
